refactor cube reps into polycubes module
