+++ rust/src/lib.rs
@@ -2,12 +2,9 @@
 mod test;
 
 pub mod iterator;
-pub mod pcube;
+pub mod polycubes;
 
-pub mod naive_polycube;
 
 pub mod hashless;
 pub mod pointlist;
-pub mod polycube_reps;
 pub mod rotation_reduced;
-pub mod rotations;
