+++ rust/src/rotation_reduced.rs
@@ -1,26 +1,12 @@
-#[cfg(feature = "diagnostics")]
-#[cfg(not(feature = "smallset"))]
-use std::collections::HashMap;
+
 use std::{cmp::max, collections::HashSet, time::Instant};
 
 use indicatif::ProgressBar;
 
-use crate::{
-    polycube_reps::CubeMap,
-    rotations::{self, rot_matrix, MatrixCol},
-};
+use crate::polycubes::{rotation_reduced::{CubeMap, rotate::{self, MatrixCol, rot_matrix}}, point_list::CubeMapPos};
 
-#[cfg(feature = "diagnostics")]
-#[cfg(feature = "size16")]
-pub static MAX_X: usize = 16;
-
-#[cfg(feature = "diagnostics")]
-#[cfg(not(feature = "size16"))]
-pub static MAX_X: usize = 32;
-
-#[cfg(feature = "smallset")]
 ///converts a cube map to  map pos for hashset storage slow (+10% runtime combined with decode last measured)
-fn cube_map_to_cube_map_pos(map: &CubeMap) -> CubeMapPos {
+fn cube_map_to_cube_map_pos(map: &CubeMap) -> CubeMapPos<16> {
     let mut pos = CubeMapPos { cubes: [0; 16] };
     let mut i = 0;
     for z in 0..=map.z as usize {
@@ -34,19 +20,11 @@ fn cube_map_to_cube_map_pos(map: &CubeMap) -> CubeMapPos {
         }
     }
     pos.cubes[i - 1] |= 0x8000;
-    #[cfg(feature = "diagnostics")]
-    {
-        let a = cube_map_from_cube_map_pos(&pos);
-        if a != *map {
-            panic!("{:?} {:?} unequal", a, map);
-        }
-    }
     pos
 }
 
-#[cfg(feature = "smallset")]
 ///converts a mappos from hashset storage to a cube map
-fn cube_map_from_cube_map_pos(map: &CubeMapPos) -> CubeMap {
+fn cube_map_from_cube_map_pos(map: &CubeMapPos<16>) -> CubeMap {
     let mut dst = CubeMap {
         x: 0,
         y: 0,
@@ -81,31 +59,12 @@ fn cube_map_from_cube_map_pos(map: &CubeMapPos) -> CubeMap {
     dst
 }
 
-#[cfg(feature = "smallset")]
-type CubeEncoding = CubeMapPos;
-#[cfg(not(feature = "smallset"))]
-type CubeEncoding = CubeMap;
-
 #[inline]
 fn insert_map(
     map: &CubeMap,
-    seen: &mut HashSet<CubeEncoding>,
-    #[cfg(feature = "diagnostics")] depth: usize,
+    seen: &mut HashSet<CubeMapPos<16>>
 ) {
-    let work_map = rotations::to_min_rot(map);
-    #[cfg(feature = "diagnostics")]
-    {
-        if map.count_cubes() != depth {
-            panic!("{:?} doesnt have {} cubes", map, depth)
-        }
-        if !map.validate_bounds() {
-            panic!("{:?} has blocks out of bounds", map)
-        }
-        if !map.validate_continuity() {
-            panic!("{:?} non continuous", map)
-        }
-    }
-    #[cfg(feature = "smallset")]
+    let work_map = rotate::to_min_rot(map);
     let work_map = cube_map_to_cube_map_pos(&work_map);
     seen.insert(work_map);
 }
@@ -209,8 +168,7 @@ fn expand_cube_map_out(map: &CubeMap, y: usize, z: usize, offset: u32) -> CubeMa
 #[inline]
 fn expand_xs(
     map: &CubeMap,
-    seen: &mut HashSet<CubeEncoding>,
-    #[cfg(feature = "diagnostics")] depth: usize,
+    seen: &mut HashSet<CubeMapPos<16>>,
 ) {
     for yz in 0..(((map.y + 1) * (map.z + 1)) as usize) {
         let left_bits = ((map.cube_map[yz] << 1) | map.cube_map[yz]) ^ map.cube_map[yz];
@@ -221,8 +179,6 @@ fn expand_xs(
                 insert_map(
                     &expand_cube_map_left(map, yz, xoff),
                     seen,
-                    #[cfg(feature = "diagnostics")]
-                    depth,
                 );
             }
         }
@@ -231,8 +187,6 @@ fn expand_xs(
                 insert_map(
                     &expand_cube_map_right(map, yz, xoff),
                     seen,
-                    #[cfg(feature = "diagnostics")]
-                    depth,
                 );
             }
         }
@@ -243,8 +197,7 @@ fn expand_xs(
 #[inline]
 fn expand_ys(
     map: &CubeMap,
-    seen: &mut HashSet<CubeEncoding>,
-    #[cfg(feature = "diagnostics")] depth: usize,
+    seen: &mut HashSet<CubeMapPos<16>>,
 ) {
     for z in 0..=map.z as usize {
         for y in 0..=map.y as usize {
@@ -266,16 +219,12 @@ fn expand_ys(
                     insert_map(
                         &expand_cube_map_up(map, y + 1, z, xoff),
                         seen,
-                        #[cfg(feature = "diagnostics")]
-                        depth,
                     );
                 }
                 if down_bits & (1 << xoff) != 0 {
                     insert_map(
                         &expand_cube_map_down(map, y, z, xoff),
                         seen,
-                        #[cfg(feature = "diagnostics")]
-                        depth,
                     );
                 }
             }
@@ -287,8 +236,7 @@ fn expand_ys(
 #[inline]
 fn expand_zs(
     map: &CubeMap,
-    seen: &mut HashSet<CubeEncoding>,
-    #[cfg(feature = "diagnostics")] depth: usize,
+    seen: &mut HashSet<CubeMapPos<16>>,
 ) {
     for z in 0..=map.z as usize {
         for y in 0..=map.y as usize {
@@ -305,16 +253,12 @@ fn expand_zs(
                     insert_map(
                         &expand_cube_map_in(map, y, z + 1, xoff),
                         seen,
-                        #[cfg(feature = "diagnostics")]
-                        depth,
                     );
                 }
                 if out_bits & (1 << xoff) != 0 {
                     insert_map(
                         &expand_cube_map_out(map, y, z, xoff),
                         seen,
-                        #[cfg(feature = "diagnostics")]
-                        depth,
                     );
                 }
             }
@@ -326,29 +270,22 @@ fn expand_zs(
 #[inline]
 fn do_cube_expansion(
     map: &CubeMap,
-    seen: &mut HashSet<CubeEncoding>,
-    #[cfg(feature = "diagnostics")] depth: usize,
+    seen: &mut HashSet<CubeMapPos<16>>,
 ) {
     expand_xs(
         map,
         seen,
-        #[cfg(feature = "diagnostics")]
-        depth,
     );
     if map.y < map.x {
         expand_ys(
             map,
             seen,
-            #[cfg(feature = "diagnostics")]
-            depth,
         );
     }
     if map.z < map.y {
         expand_zs(
             map,
             seen,
-            #[cfg(feature = "diagnostics")]
-            depth,
         );
     }
 }
@@ -357,19 +294,12 @@ fn do_cube_expansion(
 #[inline]
 fn expand_cube_map(
     map: &CubeMap,
-    seen: &mut HashSet<CubeEncoding>,
-    #[cfg(feature = "diagnostics")] depth: usize,
+    seen: &mut HashSet<CubeMapPos<16>>,
 ) {
     do_cube_expansion(
         map,
         seen,
-        #[cfg(feature = "diagnostics")]
-        depth,
     );
-    #[cfg(feature = "diagnostics")]
-    if map.count_cubes() != depth - 1 {
-        panic!("{:?} doesnt have {} cubes", map, depth - 1)
-    }
     if map.x == map.y && map.x > 0 {
         let mut rot = CubeMap {
             x: map.x,
@@ -378,15 +308,9 @@ fn expand_cube_map(
             cube_map: [0; 36],
         };
         rot_matrix(map, &mut rot, MatrixCol::YN, MatrixCol::XN, MatrixCol::ZN);
-        #[cfg(feature = "diagnostics")]
-        if rot.count_cubes() != depth - 1 {
-            panic!("{:?} doesnt have {} cubes", rot, depth - 1)
-        }
         do_cube_expansion(
             &rot,
             seen,
-            #[cfg(feature = "diagnostics")]
-            depth,
         );
     }
     if map.y == map.z && map.y > 0 {
@@ -397,15 +321,9 @@ fn expand_cube_map(
             cube_map: [0; 36],
         };
         rot_matrix(map, &mut rot, MatrixCol::XN, MatrixCol::ZP, MatrixCol::YP);
-        #[cfg(feature = "diagnostics")]
-        if rot.count_cubes() != depth - 1 {
-            panic!("{:?} doesnt have {} cubes", rot, depth - 1)
-        }
         do_cube_expansion(
             &rot,
             seen,
-            #[cfg(feature = "diagnostics")]
-            depth,
         );
     }
     if map.x == map.z && map.x > 0 {
@@ -416,45 +334,25 @@ fn expand_cube_map(
             cube_map: [0; 36],
         };
         rot_matrix(map, &mut rot, MatrixCol::ZP, MatrixCol::YP, MatrixCol::XN);
-        #[cfg(feature = "diagnostics")]
-        if rot.count_cubes() != depth - 1 {
-            panic!("{:?} doesnt have {} cubes", rot, depth - 1)
-        }
         do_cube_expansion(
             &rot,
             seen,
-            #[cfg(feature = "diagnostics")]
-            depth,
         );
-    }
-}
-
-#[cfg(feature = "diagnostics")]
-#[cfg(not(feature = "smallset"))]
-fn to_dim(cm: &CubeMap) -> Dim {
-    Dim {
-        x: cm.x as usize + 1,
-        y: cm.y as usize + 1,
-        z: cm.z as usize + 1,
     }
 }
 
 /// expand all polycubes in set n-1 to get set n
 fn expand_cube_set(
-    in_set: &HashSet<CubeEncoding>,
-    #[cfg(feature = "diagnostics")] depth: usize,
-    out_set: &mut HashSet<CubeEncoding>,
+    in_set: &HashSet<CubeMapPos<16>>,
+    out_set: &mut HashSet<CubeMapPos<16>>,
     bar: &ProgressBar,
 ) {
     let mut i = 0;
     for map in in_set.iter() {
-        #[cfg(feature = "smallset")]
         let map = &cube_map_from_cube_map_pos(map);
         expand_cube_map(
             map,
             out_set,
-            #[cfg(feature = "diagnostics")]
-            depth,
         );
         i += 1;
         if i == 100 {
@@ -463,27 +361,6 @@ fn expand_cube_set(
         }
     }
     bar.inc(i);
-    #[cfg(feature = "diagnostics")]
-    #[cfg(not(feature = "smallset"))]
-    {
-        let mut shape_map = HashMap::new();
-        for map in out_set.iter() {
-            if map.count_cubes() != depth {
-                panic!("{:?} doesnt have {} cubes", map, depth)
-            }
-            let dim = to_dim(map);
-            shape_map.insert(
-                dim,
-                match shape_map.get(&dim) {
-                    Some(res) => res + 1,
-                    None => 1,
-                },
-            );
-        }
-        for (s, map) in shape_map.iter() {
-            println!("{}, {}, {} -> {:?}", s.x, s.y, s.z, map);
-        }
-    }
 }
 
 pub fn gen_polycubes(n: usize, bar: &ProgressBar) -> usize {
@@ -502,20 +379,15 @@ pub fn gen_polycubes(n: usize, bar: &ProgressBar) -> usize {
     insert_map(
         &unit_cube,
         &mut seeds,
-        #[cfg(feature = "diagnostics")]
-        2,
     );
     for i in 3..=n as usize {
         bar.set_message(format!("seed subsets expanded for N = {}...", i));
         expand_cube_set(
             &seeds,
-            #[cfg(feature = "diagnostics")]
-            i,
             &mut dst,
             bar,
         );
-        //if diagnostics enabled panic if the returned values are wrong
-        #[cfg(feature = "diagnostics")]
+        // panic if the returned values are wrong
         if i == 3 && dst.len() != 2 {
             panic!("{} supposed to have {} elems not {}", i, 2, dst.len())
         } else if i == 4 && dst.len() != 8 {
